Dependencies Added -> Paging 3.0 for paging the Api Requests;
DataSource created to match the Paging Pattern and create a DataSource<Any, Any> to handle the Api Requests onLoad;

1 Adapter Created (LoadStateAdapter) to handle the load state and display it on a progress bar with TextView to display the problem that happened with the request, and a retry button to fetch data again;

The old adapter was replaced with PagingDataAdapter that handles the data that is fetch from API without having to handle the onScroll method on View (Activity);

RepoRepository interface created to serve as model to RepoRepositoryImpl methods;

Usecase refactored to match the new app flow;

Using Flow to load data and put into adapter;

Lifecycle coroutine is called on Activity to load the data and put into adapter. A job is used to cancel this Coroutine before a new call to prevent any leaks;

Application package on Manifest changed from "githubrepos" to "app" to remove the warning of Manifest;

Renamed RepositoryCatalog and others to match the "Repo%" pattern in order to better readability;

App is working on Activity to load the paginated Repositories.