Dependencies Added -> Paging 3.0 for paging the Api Requests;
Started to work with Caching.
App now is cached using Room Database. To achieve this feature, the app was refactored -> Replaced the DataSource with a RemoteMediator, that helps to refresh the data and insert into database.
The repository now has the refreshData function that returns the updated data that is submitted to adapter on Activity.
Created a LoadStateAdapter to handle the loading state and it trigger two options: OnRetry button when one error happen, or submit the new data displaying a Progress Bar until the data is loaded.
Repositories stored in database have now keys associated with them(another table created), so it´s easier to query the data from database, and continue fetching when the app is closed and opened again.
Renamed some files to match the pattern "Repo%".